remove id as required field in site plugin (#702)
Make id optional in site plugin api so api calls to create a site only need the name not the id of the site being created and associations need the siteId and hostID not the id of the association.
Add hooks to the site plugin to allow for adding associations for groups and hosts via the group and host api paths.